Add coord dtype normalization

+++ o3skim/loadfunctions_tco3/SBUV_GSFC.py
@@ -43,8 +43,8 @@ def load_tco3(model_path):
     for head, chunk in tables:
         header = head[0:-2].split(" ")  # Split head strings
         year = int(header[0])  # First value in header is the year
-        table = pd.read_table(chunk, sep=DELIMITER, index_col=[0, 1])
-        lat = [(l1 + l2) / 2 for l1, l2 in table.index]
+        table = pd.read_table(chunk, sep=DELIMITER, index_col=[0, 1], dtype='float32')
+        lat = np.float32([(l1 + l2) / 2 for l1, l2 in table.index])
         time = [pd.datetime(year, m, 1) for m in range(1, 13)]
         array = xr.DataArray(table, dict(lat=lat, time=time), ["lat", "time"])
         array.values.flat[array.values.flat == 0.0] = np.nan
@@ -68,7 +68,7 @@ def load_tco3(model_path):
         },
         coords={
             "lon": xr.Variable(
-                *[["lon"], [0]],
+                *[["lon"], np.float32([0])],
                 attrs=dict(
                     standard_name="longitude",
                     axis="X",
